feat: Add customizable PaymentMethodButton widget for selecting payment methods
feat: Implement PaymentStatusCard widget to display payment transaction status

feat: Create SdkAvailabilityIndicator widget to show native SDK availability status

feat: Develop MoamalatPaymentUnified widget for unified payment processing via SDK and WebView

feat: Introduce MoamalatSdkService for handling native SDK payment transactions

feat: Define PaymentMethod enum for selecting payment processing methods

docs: Enhance documentation for payment method selection and usage examples

+++ moamalat_payment/lib/src/sucsses_model.dart
@@ -4,7 +4,7 @@ class TransactionSucsses {
   String? systemReference;
   String? networkReference;
   String? merchantReference;
-  int? amount;
+  double? amount;
   String? currency;
   String? paidThrough;
   String? payerAccount;
@@ -38,7 +38,7 @@ class TransactionSucsses {
       systemReference: json['SystemReference'],
       networkReference: json['NetworkReference'],
       merchantReference: json['MerchantReference'],
-      amount: int.parse(json['Amount']),
+      amount: double.parse(json['Amount']),
       currency: json['Currency'],
       paidThrough: json['PaidThrough'],
       payerAccount: json['PayerAccount'],
